Fix interacting with GitHub Container Registry
The GitHub container registry (ghcr.io) returns a bearer token
that doesn't have an explicit expiration time.

Prior to this change, because of this anomaly, the token was not added into the
JWT token cache. That caused all the request done against ghcr.io to
be unauthenticated. That broke any kind of interaction with ghcr, even with
public images.

This code handles the lack of an expiration claim by setting a
60 second expiration window.
According to the official documentation (https://docs.docker.com/registry/spec/auth/token/#requesting-a-token),
this behaviour is fine:

> The duration in seconds since the token was issued that it will remain valid.
> When omitted, this defaults to 60 seconds. For compatibility with older clients,
> a token should never be returned with less than 60 seconds to live.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>